refactor: разделить взаимодействие с базой данных
