+++ README.md
@@ -1,2 +1,3 @@
 # java-filmorate
-Template repository for Filmorate project.
+Cсылка на файл диаграммы:
+https://dbdiagram.io/d/64382d6e8615191cfa8dba43
