+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/GenreDbStorage.java
@@ -1,0 +1,42 @@
+package ru.yandex.practicum.filmorate.storage.film;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Repository;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.model.Genre;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.List;
+
+@Repository
+@Slf4j
+public class GenreDbStorage {
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public GenreDbStorage(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public Genre getGenreById(int id) {
+        try {
+            String sqlQuery = "select* from GENRE where GENRE_ID = ?";
+            return jdbcTemplate.queryForObject(sqlQuery, this::mapRowToGenre, id);
+        } catch (EmptyResultDataAccessException e) {
+            log.warn("Ошибка валидации наличия жанра");
+            throw new NotFoundException("Жанр с таким айди отсутствует");
+        }
+    }
+
+    public List<Genre> getGenres() {
+        String sqlQuery = "select* from GENRE";
+        return jdbcTemplate.query(sqlQuery, this::mapRowToGenre);
+    }
+
+    private Genre mapRowToGenre(ResultSet resultSet, int rowNum) throws SQLException {
+        return new Genre(resultSet.getInt("GENRE_ID"), resultSet.getString("NAME"));
+    }
+}
