+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/MpaDbStorage.java
@@ -1,0 +1,42 @@
+package ru.yandex.practicum.filmorate.storage.film;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Repository;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.model.Mpa;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.List;
+
+@Repository
+@Slf4j
+public class MpaDbStorage {
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public MpaDbStorage(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public Mpa getMpaById(int id) {
+        try {
+            String sqlQuery = "select* from RATING where RATING_ID = ?";
+            return jdbcTemplate.queryForObject(sqlQuery, this::mapRowToMpa, id);
+        } catch (EmptyResultDataAccessException e) {
+            log.warn("Ошибка валидации наличия рейтинга");
+            throw new NotFoundException("Рейтинг с таким айди отсутствует");
+        }
+    }
+
+    public List<Mpa> getMpas() {
+        String sqlQuery = "select* from RATING";
+        return jdbcTemplate.query(sqlQuery, this::mapRowToMpa);
+    }
+
+    private Mpa mapRowToMpa(ResultSet resultSet, int rowNum) throws SQLException {
+        return new Mpa(resultSet.getInt("RATING_ID"), resultSet.getString("RATING_NAME"));
+    }
+}
